core/ test follow passing
