+++ core/drand_test.go
@@ -2,17 +2,13 @@ package core
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	gnet "net"
 	"os"
 	"path"
-	"sync"
 	"testing"
 	"time"
-
-	"google.golang.org/grpc"
 
 	"github.com/drand/drand/chain"
 	"github.com/drand/drand/key"
@@ -21,7 +17,6 @@ import (
 	"github.com/drand/drand/protobuf/drand"
 	"github.com/drand/drand/test"
 
-	clock "github.com/jonboulle/clockwork"
 	"github.com/kabukky/httpscerts"
 	"github.com/stretchr/testify/require"
 )
@@ -220,407 +215,6 @@ func TestDrandResharePreempt(t *testing.T) {
 	dt.TestPublicBeacon(dt.Ids(1, false)[0], false)
 }
 
-type Node struct {
-	addr  string
-	drand *Drand
-	clock clock.FakeClock
-}
-
-type DrandTest2 struct {
-	sync.Mutex
-	t *testing.T
-	// tmp dir for certificates, keys etc
-	dir          string
-	newDir       string
-	certPaths    []string
-	newCertPaths []string
-	// global clock on which all drand's clock are synchronized
-	clock clock.FakeClock
-
-	n      int
-	thr    int
-	newN   int
-	newThr int
-	period time.Duration
-	// only set after the DKG
-	group *key.Group
-	// needed to give the group to new nodes during a resharing - only set after
-	// a successful DKG
-	groupPath string
-	// only set after the resharing
-	newGroup *key.Group
-	// nodes that are created for running a first DKG
-	nodes []*Node
-	// new additional nodes that are created for running a resharing
-	newNodes []*Node
-	// nodes that actually ran the resharing phase - it's a combination of nodes
-	// and new nodes. These are the one that should appear in the newGroup
-	resharedNodes []*Node
-}
-
-// NewDrandTest creates a drand test scenario with initial n nodes and ready to
-// run a DKG for the given threshold that will then launch the beacon with the
-// specified period
-func NewDrandTest2(t *testing.T, n, thr int, period time.Duration) *DrandTest2 {
-	dt := new(DrandTest2)
-	drands, _, dir, certPaths := BatchNewDrand(n, false,
-		WithCallOption(grpc.WaitForReady(true)),
-	)
-	dt.t = t
-	dt.dir = dir
-	dt.certPaths = certPaths
-	dt.n = n
-	dt.thr = thr
-	dt.period = period
-	dt.clock = clock.NewFakeClock()
-	dt.nodes = make([]*Node, 0, n)
-	for _, drand := range drands {
-		node := dt.newNode(drand)
-		dt.nodes = append(dt.nodes, node)
-	}
-	dt.groupPath = path.Join(dt.dir, "group.toml")
-	return dt
-}
-
-// Ids returns the list of the first n ids given the newGroup parameter (either
-// in the original group or the reshared one)
-func (d *DrandTest2) Ids(n int, newGroup bool) []string {
-	nodes := d.nodes
-	if newGroup {
-		nodes = d.resharedNodes
-	}
-	var ids []string
-	for _, node := range nodes[:n] {
-		ids = append(ids, node.addr)
-	}
-	return ids
-}
-
-// RunDKG runs the DKG with the initial nodes created during NewDrandTest
-func (d *DrandTest2) RunDKG() *key.Group {
-	// common secret between participants
-	secret := "thisisdkg"
-	root := d.nodes[0]
-	controlClient, err := net.NewControlClient(root.drand.opts.controlPort)
-	require.NoError(d.t, err)
-	// the root node will return the group over this channel
-	var wg sync.WaitGroup
-	wg.Add(d.n)
-	// first run the leader and then run the other nodes
-	go func() {
-		_, err := controlClient.InitDKGLeader(d.n, d.thr, d.period, testDkgTimeout, nil, secret, testBeaconOffset)
-		require.NoError(d.t, err)
-		fmt.Printf("\n\nTEST LEADER FINISHED\n\n")
-		wg.Done()
-	}()
-
-	// make sure the leader is up and running to start the setup
-	time.Sleep(1 * time.Second)
-	// all other nodes will send their PK to the leader that will create the
-	// group
-	for _, node := range d.nodes[1:] {
-		go func(n *Node) {
-			client, err := net.NewControlClient(n.drand.opts.controlPort)
-			require.NoError(d.t, err)
-			_, err = client.InitDKG(root.drand.priv.Public, nil, secret)
-			fmt.Printf("\n\nTEST NONLEADER FINISHED\n\n")
-			require.NoError(d.t, err)
-			wg.Done()
-			fmt.Printf("\n\n\n TESTDKG NON-ROOT %s FINISHED\n\n\n", n.addr)
-		}(node)
-	}
-
-	// wait for all to return
-	wg.Wait()
-	fmt.Printf("\n\n\n TESTDKG ROOT %s FINISHED\n\n\n", root.addr)
-	// we check that we can fetch the group using control functionalities on the root node
-	groupProto, err := controlClient.GroupFile()
-	require.NoError(d.t, err)
-	group, err := key.GroupFromProto(groupProto)
-	require.NoError(d.t, err)
-	// we check all nodes are included in the group
-	for _, node := range d.nodes {
-		require.NotNil(d.t, group.Find(node.drand.priv.Public))
-	}
-	// we check the group has the right threshold
-	require.Len(d.t, group.PublicKey.Coefficients, d.thr)
-	require.Equal(d.t, d.thr, group.Threshold)
-	require.NoError(d.t, key.Save(d.groupPath, group, false))
-	d.group = group
-	fmt.Println("setup dkg group:", d.group.String())
-	return group
-}
-
-func (d *DrandTest2) Cleanup() {
-	os.RemoveAll(d.dir)
-	os.RemoveAll(d.newDir)
-}
-
-// GetBeacon returns the beacon of the given round for the specified drand id
-func (d *DrandTest2) GetBeacon(id string, round int, newGroup bool) (*chain.Beacon, error) {
-	nodes := d.nodes
-	if newGroup {
-		nodes = d.resharedNodes
-	}
-	for _, node := range nodes {
-		if node.addr != id {
-			continue
-		}
-		return node.drand.beacon.Store().Get(uint64(round))
-	}
-	return nil, errors.New("that should not happen")
-}
-
-// GetDrand returns the node associated with this address, either in the new
-// group or the current group.
-func (d *DrandTest2) GetDrand(id string, newGroup bool) *Node {
-	nodes := d.nodes
-	if newGroup {
-		nodes = d.resharedNodes
-	}
-	for _, node := range nodes {
-		if node.addr == id {
-			return node
-		}
-	}
-	panic("no nodes found at this id")
-}
-
-// StopDrand stops a node from the first group
-func (d *DrandTest2) StopDrand(id string, newGroup bool) {
-	node := d.GetDrand(id, newGroup)
-	dr := node.drand
-	dr.Stop(context.Background())
-	pinger, err := net.NewControlClient(dr.opts.controlPort)
-	require.NoError(d.t, err)
-	var counter = 1
-	fmt.Println(" DRAND ", dr.priv.Public.Address(), " TRYING TO PING")
-	for range time.Tick(100 * time.Millisecond) {
-		if err := pinger.Ping(); err != nil {
-			fmt.Println(" DRAND ", dr.priv.Public.Address(), " TRYING TO PING DONE")
-			break
-		}
-		counter++
-		require.LessOrEqual(d.t, counter, 5)
-	}
-	fmt.Println(" DRAND ", dr.priv.Public.Address(), " STOPPED")
-}
-
-// StartDrand fetches the drand given the id, in the respective group given the
-// newGroup parameter and runs the beacon
-func (d *DrandTest2) StartDrand(id string, catchup, newGroup bool) {
-	node := d.GetDrand(id, newGroup)
-	dr := node.drand
-	// we load it from scratch as if the operator restarted its node
-	newDrand, err := LoadDrand(dr.store, dr.opts)
-	require.NoError(d.t, err)
-	node.drand = newDrand
-	fmt.Println("--- JUST BEFORE STARTBEACON---")
-	newDrand.StartBeacon(catchup)
-	fmt.Println("--- JUST AFTER STARTBEACON---")
-}
-
-func (d *DrandTest2) Now() time.Time {
-	return d.clock.Now()
-}
-
-// MoveToTime sets the clock of all drands to the designated unix timestamp in
-// seconds
-func (d *DrandTest2) MoveToTime(target int64) {
-	now := d.Now().Unix()
-	if now < target {
-		d.MoveTime(time.Duration(target-now) * time.Second)
-	} else {
-		fmt.Println("ALREADY PASSED")
-	}
-}
-
-// MoveTime advance the clock of all drand of the given duration
-func (d *DrandTest2) MoveTime(p time.Duration) {
-	for _, node := range d.nodes {
-		node.clock.Advance(p)
-	}
-	for _, node := range d.newNodes {
-		node.clock.Advance(p)
-	}
-	d.clock.Advance(p)
-	fmt.Printf(" --- MoveTime: new time is %d \n", d.Now().Unix())
-	time.Sleep(getSleepDuration())
-}
-
-// TestBeaconLength looks if the beacon chain on the given ids is of the
-// expected length
-func (d *DrandTest2) TestBeaconLength(length int, newGroup bool, ids ...string) {
-	nodes := d.nodes
-	if newGroup {
-		nodes = d.resharedNodes
-	}
-	for _, id := range ids {
-		for _, node := range nodes {
-			if node.addr != id {
-				continue
-			}
-			var found bool
-			for i := 0; i < 10; i++ {
-				inst := node.drand
-				if length != inst.beacon.Store().Len() {
-					time.Sleep(getSleepDuration())
-					continue
-				}
-				found = true
-				break
-			}
-			require.True(d.t, found, "node %d not have enough beacon", id)
-		}
-	}
-}
-
-// TestPublicBeacon looks if we can get the latest beacon on this node
-func (d *DrandTest2) TestPublicBeacon(id string, newGroup bool) *drand.PublicRandResponse {
-	node := d.GetDrand(id, newGroup)
-	dr := node.drand
-	client := net.NewGrpcClientFromCertManager(dr.opts.certmanager, dr.opts.grpcOpts...)
-	resp, err := client.PublicRand(context.TODO(), test.NewTLSPeer(dr.priv.Public.Addr), &drand.PublicRandRequest{})
-	require.NoError(d.t, err)
-	require.NotNil(d.t, resp)
-	return resp
-}
-
-// SetupNewNodes creates new additional nodes that can participate during the
-// resharing
-func (d *DrandTest2) SetupNewNodes(newNodes int) {
-	newDrands, _, newDir, newCertPaths := BatchNewDrand(newNodes, false,
-		WithCallOption(grpc.WaitForReady(false)), WithLogLevel(log.LogDebug))
-	d.newCertPaths = newCertPaths
-	d.newDir = newDir
-	d.newNodes = make([]*Node, newNodes)
-	// add certificates of new nodes to the old nodes
-	for _, node := range d.nodes {
-		inst := node.drand
-		for _, cp := range newCertPaths {
-			inst.opts.certmanager.Add(cp)
-		}
-	}
-	// store new part. and add certificate path of current nodes to the new
-	d.newNodes = make([]*Node, 0, newNodes)
-	for _, inst := range newDrands {
-		node := d.newNode(inst)
-		d.newNodes = append(d.newNodes, node)
-		for _, cp := range d.certPaths {
-			inst.opts.certmanager.Add(cp)
-		}
-	}
-}
-
-// RunReshare runs the resharing procedure with only "oldRun" current nodes
-// running, and "newRun" new nodes running (the ones created via SetupNewNodes).
-// It sets the given threshold to the group.
-// It stops the nodes excluded first.
-func (d *DrandTest2) RunReshare(oldRun, newRun, newThr int, timeout time.Duration) *key.Group {
-	fmt.Printf(" -- Running RESHARE with %d/%d old, %d/%d new nodes\n", oldRun, len(d.nodes), newRun, len(d.newNodes))
-	var clientCounter = new(sync.WaitGroup)
-	var secret = "thisistheresharing"
-	total := oldRun + newRun
-	// stop the exluded nodes
-	for _, node := range d.nodes[oldRun:] {
-		d.StopDrand(node.addr, false)
-	}
-	for _, node := range d.newNodes[newRun:] {
-		d.StopDrand(node.addr, true)
-	}
-
-	d.newN = total
-	d.newThr = newThr
-	leader := d.nodes[0]
-	// function that each non-leader runs to start the resharing
-	runreshare := func(n *Node, newNode bool) {
-		dr := n.drand
-		// instruct to be ready for a reshare
-		client, err := net.NewControlClient(dr.opts.controlPort)
-		require.NoError(d.t, err)
-		_, err = client.InitReshare(leader.drand.priv.Public, secret, d.groupPath)
-		require.NoError(d.t, err, "non-leader node (new?%v) error during reshare ", newNode)
-		fmt.Printf("\n\nRESHARING TEST: non-leader drand %s DONE RESHARING - %s\n", dr.priv.Public.Address(), dr.priv.Public.Key)
-		clientCounter.Done()
-	}
-
-	clientCounter.Add(1)
-	groupCh := make(chan *key.Group, 1)
-	// first run the leader, then the other nodes will send their PK to the
-	// leader and then the leader will answer back with the new group
-	go func() {
-		oldNode := d.group.Find(leader.drand.priv.Public)
-		if oldNode == nil {
-			panic("leader not found in old group")
-		}
-		// old root: oldNode.Index leater: leader.addr
-		client, err := net.NewControlClient(leader.drand.opts.controlPort)
-		require.NoError(d.t, err)
-		finalGroup, err := client.InitReshareLeader(d.newN, d.newThr, timeout, secret, "", testBeaconOffset)
-		// Done resharing
-		if err != nil {
-			panic(err)
-		}
-		clientCounter.Done()
-		fg, err := key.GroupFromProto(finalGroup)
-		if err != nil {
-			panic(err)
-		}
-		groupCh <- fg
-	}()
-	d.resharedNodes = append(d.resharedNodes, leader)
-	// leave some time to make sure leader is listening
-	time.Sleep(1 * time.Second)
-
-	// run the current nodes
-	for _, node := range d.nodes[1:oldRun] {
-		d.resharedNodes = append(d.resharedNodes, node)
-		clientCounter.Add(1)
-		go runreshare(node, false)
-	}
-
-	// run the new ones
-	for _, node := range d.newNodes[:newRun] {
-		d.resharedNodes = append(d.resharedNodes, node)
-		fmt.Printf("\n ++ NEW NODE running RESHARE: %s\n", node.addr)
-		clientCounter.Add(1)
-		go runreshare(node, true)
-	}
-	// wait for the return of the clients
-	checkWait(clientCounter)
-	finalGroup := <-groupCh
-	d.newGroup = finalGroup
-	return finalGroup
-}
-
-// newNode creates a node struct from a drand and sets the clock according to
-// the drand test clock.
-func (d *DrandTest2) newNode(dr *Drand) *Node {
-	id := dr.priv.Public.Address()
-	now := d.clock.Now()
-	c := clock.NewFakeClockAt(now)
-	dr.opts.clock = c
-	return &Node{
-		addr:  id,
-		drand: dr,
-		clock: c,
-	}
-}
-
-func checkWait(counter *sync.WaitGroup) {
-	var doneCh = make(chan bool, 1)
-	go func() {
-		counter.Wait()
-		doneCh <- true
-	}()
-	select {
-	case <-doneCh:
-		break
-	case <-time.After(11 * time.Second):
-		panic("outdated beacon time")
-	}
-}
-
 // Check they all have same chain info
 func TestDrandPublicChainInfo(t *testing.T) {
 	n := 10
@@ -699,11 +293,14 @@ func TestDrandPublicRand(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, i, resp.Round)
 		fmt.Println("REQUEST ROUND ", i, " GOT ROUND ", resp.Round)
+
 	}
 }
 
 // Test if the we can correctly fetch the rounds after a DKG using the
 // PublicRandStream RPC call
+// It also test the follow method call (it avoid redoing an expensive and long
+// setup on CI to test both behaviors).
 func TestDrandPublicStream(t *testing.T) {
 	n := 4
 	thr := key.DefaultThreshold(n)
@@ -772,6 +369,64 @@ func TestDrandPublicStream(t *testing.T) {
 	case <-time.After(50 * time.Millisecond):
 		// correct
 	}
+}
+func TestDrandFollowChain(tt *testing.T) {
+	n := 4
+	thr := key.DefaultThreshold(n)
+	p := 1 * time.Second
+	dt := NewDrandTest2(tt, n, thr, p)
+	defer dt.Cleanup()
+	group := dt.RunDKG()
+	time.Sleep(getSleepDuration())
+	root := dt.nodes[0]
+	rootID := root.drand.priv.Public
+
+	dt.MoveToTime(group.GenesisTime)
+	// do a few periods
+	for i := 0; i < 10; i++ {
+		dt.MoveTime(group.Period)
+	}
+
+	cm := root.drand.opts.certmanager
+	client := net.NewGrpcClientFromCertManager(cm)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	// get last round first
+	resp, err := client.PublicRand(ctx, rootID, new(drand.PublicRandRequest))
+	require.NoError(tt, err)
+
+	// TEST setup a new node and fetch history
+
+	newNode := dt.SetupNewNodes(1)[0]
+	newClient, err := net.NewControlClient(newNode.drand.opts.controlPort)
+	require.NoError(tt, err)
+	addrToFollow := []string{rootID.Address()}
+	hash := fmt.Sprintf("%x", chain.NewChainInfo(group).Hash())
+	tls := true
+	progress, errCh, err := newClient.StartFollowChain(context.Background(), hash, addrToFollow, tls)
+	require.NoError(tt, err)
+	for {
+		select {
+
+		case p := <-progress:
+			if p.Current == resp.GetRound() {
+				// success
+				fmt.Printf("\n\nSUCCESSSSSS\n\n")
+				return
+			}
+		case e := <-errCh:
+			require.NoError(tt, e)
+		case <-time.After(1 * time.Second):
+			tt.FailNow()
+		}
+	}
+
+	store, err := newNode.drand.createBoltStore()
+	require.NoError(tt, err)
+	lastB, err := store.Last()
+	require.NoError(tt, err)
+	require.Equal(tt, resp.GetRound(), lastB.Round)
+
 }
 
 // Test if the we can correctly fetch the rounds through the local proxy
@@ -898,7 +553,7 @@ func CloseAllDrands(drands []*Drand) {
 func getSleepDuration() time.Duration {
 	if os.Getenv("CIRCLE_CI") != "" {
 		fmt.Println("--- Sleeping on CIRCLECI")
-		return time.Duration(1000) * time.Millisecond
+		return time.Duration(600) * time.Millisecond
 	}
 	return time.Duration(500) * time.Millisecond
 }
